src/requirements.txt
customtkinter~=5.2.2
opencv-python~=4.11.0.86
PyYAML~=6.0.2
pillow~=11.1.0
tqdm~=4.67.1
ImageHash~=4.3.1
numpy~=2.2.1
ultralytics~=8.3.65